dyper-ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.3.0
pillow==10.4.0
opencv-python-headless==4.9.0.80
ultralytics>=8.3.0
python-dotenv==1.0.0
